net-test: packetdrill: tcpdump: reduce sleep
As suggested by Davide, TCPDump's '--immediate-mode' will force the
kernel to flush packets to userspace immediately after the reception.
Thanks to that, the 'sleep 1' at the end of the test is no longer
needed: the userspace will have all the packets when the kill signal
will be received. Just to be on the safe side, we can keep a sleep 0.1,
as recommended by Willem.

It looks like this TCPDump option is usually not recommended but because
there should not be too many packets generated here with Packetdrill, it
seems fine to have it and saves 1 second for each test. Note that it is
not an exotic option as it is used by default when packets are not
written in a file. Also, we now use '--packet-buffered' to reduce IO.

Now that the "immediate mode" is being used, it seems enough to just
wait for the '.pcap' file to be created to know when TCPDump is ready.

While at it, the directory where the capture will be store is created if
it didn't exist.

Suggested-by: Davide Caratti <dcaratti@redhat.com>
Signed-off-by: Matthieu Baerts <matttbe@kernel.org>